limiter: move to internal/backend
